moved style to a style file instead
